Refatorando o código + notificações de matches no botão de ir para a tela de matches
